% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_estimators.R
\name{add_pet_peese_corrected}
\alias{add_pet_peese_corrected}
\title{Correct PET-PEESE Estimates}
\usage{
add_pet_peese_corrected(data)
}
\arguments{
\item{data}{A data frame of results, including `ma_method` and `b` columns.}
}
\value{
A data frame with added corrected rows for PET-PEESE.
}
\description{
Filters the results of PET-PEESE and sets negative effect sizes to zero.
}
