% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_estimators.R
\name{calculate_waap}
\alias{calculate_waap}
\title{Calculate WAAP Estimates}
\usage{
calculate_waap(dat)
}
\arguments{
\item{dat}{A data frame containing `yi` (effect size) and `vi` (variance).}
}
\value{
A list containing `b`, `ci.lb`, `ci.ub`, and `pval`.
}
\description{
Performs Weighted Average of Adequately Powered (WAAP) estimation.
}
