% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multiverse_report.R
\name{create_interactive_multiverse_plots}
\alias{create_interactive_multiverse_plots}
\title{Create Interactive Plotly Versions of Multiverse Plots}
\usage{
create_interactive_multiverse_plots(results, wf_columns = NULL)
}
\arguments{
\item{results}{Data frame containing multiverse analysis results}

\item{wf_columns}{Character vector of which factor column names}
}
\value{
List of interactive plotly objects
}
\description{
Create Interactive Plotly Versions of Multiverse Plots
}
