% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_estimators.R
\name{fit_hk_sj}
\alias{fit_hk_sj}
\title{Hartung-Knapp with Sidik-Jonkman}
\usage{
fit_hk_sj(data)
}
\arguments{
\item{data}{Data frame containing yi (effect sizes) and vi (variances)}
}
\value{
A universe_result object with HK/SJ estimate
}
\description{
Implements random effects meta-analysis with Hartung-Knapp confidence intervals
and Sidik-Jonkman variance estimation.
}
\details{
The Hartung-Knapp method provides improved confidence interval coverage
by using a t-distribution instead of normal distribution. Combined with
Sidik-Jonkman \eqn{\tau^2} estimation, this can provide better small-sample properties.
}
\references{
Knapp, G., & Hartung, J. (2003). Improved tests for a random effects meta-regression
with a single covariate. Statistics in Medicine, 22(17), 2693-2710.
https://doi.org/10.1002/sim.1482

Sidik, K., & Jonkman, J. N. (2005). Simple heterogeneity variance estimation for meta-analysis.
Journal of the Royal Statistical Society, Series C, 54(2), 367-384.
https://doi.org/10.1111/j.1467-9876.2005.00489.x
}
\keyword{internal}
