% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_spec_curve.R
\name{generate_dynamic_labels}
\alias{generate_dynamic_labels}
\title{Generate Dynamic Labels}
\usage{
generate_dynamic_labels(data, lookup_table)
}
\arguments{
\item{data}{A data frame containing the specification data.}

\item{lookup_table}{A named list mapping column names to human-readable labels.}
}
\value{
A named vector of human-readable labels for each relevant column.
}
\description{
Dynamically maps column names to human-readable labels using
  a lookup table.
}
