% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multiverse_report.R
\name{generate_report_text}
\alias{generate_report_text}
\title{Generate Report Text}
\usage{
generate_report_text(summary_stats, wf_analysis, wf_columns, format)
}
\description{
Generate Report Text
}
\keyword{internal}
