% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse-result-class.R
\name{new_multiverse_result}
\alias{new_multiverse_result}
\title{Build a multiverse result (internal helper)}
\usage{
new_multiverse_result(b, ci.lb, ci.ub, pval)
}
\arguments{
\item{b, ci.lb, ci.ub, pval}{Numeric scalars.}
}
\value{
An object of class \code{c("multiverse_result","list")}.
}
\description{
Build a multiverse result (internal helper)
}
\keyword{internal}
