% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{parse_advanced_factor}
\alias{parse_advanced_factor}
\title{Parse advanced factor definition}
\usage{
parse_advanced_factor(factor_def, factor_name, data)
}
\description{
Parse advanced factor definition
}
\keyword{internal}
