% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{parse_custom_groups}
\alias{parse_custom_groups}
\title{Parse custom group definitions}
\usage{
parse_custom_groups(
  factor_def,
  factor_name,
  column_name,
  available_levels,
  decision_type
)
}
\description{
Parse custom group definitions
}
\keyword{internal}
