% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{parse_simple_factor}
\alias{parse_simple_factor}
\title{Parse simple factor definition}
\usage{
parse_simple_factor(factor_def, factor_name, data)
}
\description{
Parse simple factor definition
}
\keyword{internal}
