% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{parse_special_type}
\alias{parse_special_type}
\title{Parse special type definitions (ordered, binary, etc.)}
\usage{
parse_special_type(
  factor_def,
  factor_name,
  column_name,
  available_levels,
  decision_type
)
}
\description{
Parse special type definitions (ordered, binary, etc.)
}
\keyword{internal}
