% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_VoE.R
\name{plotly_VoE}
\alias{plotly_VoE}
\title{Interactive Vibration of Effects (VoE) Plot}
\usage{
plotly_VoE(
  data,
  x = "b",
  y = "pval",
  colorblind_friendly = TRUE,
  cutoff = 10,
  x_breaks = seq(-0.5, 2, by = 0.25),
  y_breaks = c(1e-11, 1e-10, 1e-09, 1e-08, 1e-07, 1e-06, 1e-05, 1e-04, 0.001, 0.01, 0.05,
    0.1, 0.5, 1),
  x_limits = NULL,
  y_limits = c(1e-11, 1),
  vertical_lines = c(0.1, 0.9),
  hline_value = 0.05,
  title_template = "{k} meta-analyses with at least {cutoff} studies.",
  interactive = TRUE
)
}
\arguments{
\item{data}{Data frame containing the input data for plotting.}

\item{x}{Character. Name of the x-variable column (default: "b").}

\item{y}{Character. Name of the y-variable column (default: "pvalue").}

\item{colorblind_friendly}{Logical. Use a colorblind-friendly palette (default: TRUE).}

\item{cutoff}{Numeric. Minimum number of studies for inclusion (default: 10).}

\item{x_breaks}{Numeric vector. Break points for the x-axis.}

\item{y_breaks}{Numeric vector. Break points for the y-axis.}

\item{x_limits}{Numeric vector. Limits for the x-axis.}

\item{y_limits}{Numeric vector. Limits for the y-axis.}

\item{vertical_lines}{Numeric vector. Quantiles for vertical reference lines (default: c(0.1, 0.9)).}

\item{hline_value}{Numeric. Horizontal reference line value (default: 0.05).}

\item{title_template}{Character. Template for the plot title (supports glue syntax).}

\item{interactive}{Logical. Whether to create an interactive Plotly plot (`TRUE`) or static ggplot plot (`FALSE`).}
}
\value{
A `plotly` object containing the interactive vibration of effects plot if `interactive = TRUE`, or a combined `ggplot` object if `interactive = FALSE`.
}
\description{
Creates an interactive plot visualizing the vibration of effects with effect sizes and p-values.
}
\examples{
plotly_VoE(data, x = "b", y = "pval", cutoff = 10)

}
