% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{print_factor_summary}
\alias{print_factor_summary}
\title{Print user-friendly summary of factor setup}
\usage{
print_factor_summary(factor_info, factor_groups, data)
}
\description{
Print user-friendly summary of factor setup
}
\keyword{internal}
