% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{process_custom_factor}
\alias{process_custom_factor}
\title{Process custom factor (internal)}
\usage{
process_custom_factor(factor_def, factor_name, data, index)
}
\description{
Process custom factor (internal)
}
\keyword{internal}
