% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_factors.R
\name{process_simple_factor}
\alias{process_simple_factor}
\title{Process simple factor (internal)}
\usage{
process_simple_factor(column_def, factor_name, data, index)
}
\description{
Process simple factor (internal)
}
\keyword{internal}
