% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid.R
\name{summary.mv_spec_grid}
\alias{summary.mv_spec_grid}
\title{Summarize a specification grid}
\usage{
\method{summary}{mv_spec_grid}(object, ...)
}
\arguments{
\item{object}{An \code{mv_spec_grid}.}

\item{...}{Unused.}
}
\description{
Summarize a specification grid
}
