% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_estimators.R
\name{fit_pm}
\alias{fit_pm}
\title{Paule-Mandel Random Effects}
\usage{
fit_pm(data)
}
\arguments{
\item{data}{Data frame containing yi (effect sizes) and vi (variances)}
}
\value{
A universe_result object with Paule-Mandel estimate
}
\description{
Implements random effects meta-analysis using the Paule-Mandel method
for between-study variance estimation.
}
\details{
The Paule-Mandel method estimates τ² by equating the weighted sum of squares
to its expected value. It can perform better than DerSimonian-Laird in
certain conditions and provides an exact solution.
}
\references{
Paule, R. C., & Mandel, J. (1982). Consensus values and weighting factors.
Journal of Research of the National Bureau of Standards, 87(5), 377-385.
https://doi.org/10.6028/jres.087.022
}
\keyword{internal}
