% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_estimators.R
\name{fit_uwls}
\alias{fit_uwls}
\title{Fit UWLS Estimates}
\usage{
fit_uwls(dat)
}
\arguments{
\item{dat}{A data frame containing `yi` (effect size) and `vi` (variance).}
}
\value{
A list containing `b`, `ci.lb`, `ci.ub`, and `pval`.
}
\description{
Performs Unweighted Least Squares (UWLS) estimation.
}
