% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_multiverse.R
\name{general_multiverse}
\alias{general_multiverse}
\title{General Multiverse Meta-Analysis Handler}
\usage{
general_multiverse(i, data_multiverse, specifications, k_smallest_ma = 5)
}
\arguments{
\item{i}{Integer. The row index in the `specifications` data frame, representing a single
specification to be applied to the data.}

\item{data_multiverse}{Data frame. The dataset containing all potential studies and factors
for the multiverse analysis.}

\item{specifications}{Data frame. Contains all specifications generated using
\code{\link{create_multiverse_specifications}}. Includes columns for "Which" and "How" factors.}

\item{k_smallest_ma}{Integer. The smallest number of unique studies required for a valid meta-analysis.
Defaults to 5.}
}
\value{
A data frame containing the results of the specified meta-analysis, including the:
  - Effect size estimate (\code{b}),
  - Confidence intervals (\code{ci.lb} and \code{ci.ub}),
  - P-value (\code{pval}),
  - Number of studies included in the subset (\code{k}),
  - Set of study IDs included in the analysis (\code{set}),
  - The corresponding specification parameters.
}
\description{
A core function for dynamically applying multiverse meta-analysis specifications to subsets
of data, applying "Which" factors and "How" factors based on user-specified configurations.
}
\details{
This function dynamically filters the input data based on "Which" factors (e.g.,
\code{wf_1}, \code{wf_2}, etc.) and applies "How" factors (\code{ma_method} and
\code{dependency}). The analysis only proceeds if the subset contains at least
\code{k_smallest_ma} unique studies. Supports both \code{aggregate} and \code{modeled} dependencies.
}
\examples{
# Example of applying a specification:
data_multiverse <- data.frame(
  study = rep(paste0("S", 1:5), each = 2),
  es_id = 1:10,
  yi = rnorm(10, mean = 0.5),
  vi = runif(10, min = 0.01, max = 0.1),
  wf_1 = rep(c("A", "B"), times = 5)
  )
specifications <- data.frame(
  wf_1 = c("A", "B"),
  dependency = "aggregate",
  ma_method = "reml"
)
result <- general_multiverse(1,
   data_multiverse,
   specifications)

}
\seealso{
\code{\link{create_multiverse_specifications}}, \code{\link{run_aggregate_dependency}},
  \code{\link{run_modeled_dependency}}
}
