% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_voe.R
\name{generate_tooltip_voe}
\alias{generate_tooltip_voe}
\title{Generate Tooltips for Vibration of Effects Plot}
\usage{
generate_tooltip_voe(data, x, y, k_col = "k", set_col = "set")
}
\arguments{
\item{data}{Data frame containing the input data.}

\item{x}{Character. Name of the x-variable column for the effect size}

\item{y}{Character. Name of the y-variable column for the pvalue.}

\item{k_col}{Character. Name of the column containing the number of studies (default: "k").}

\item{set_col}{Character. Name of the column containing the study set IDs (default: "set").}
}
\value{
A data frame with an additional `tooltip` column.
}
\description{
Creates a `tooltip` column for interactive plots by dynamically summarizing key information.
}
\examples{
generate_tooltip(data, x = "b", y = "pvalue")

}
