% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-multiverse.R
\name{get_original_names}
\alias{get_original_names}
\title{Get Original Column Names}
\usage{
get_original_names(wf_mapping, wf_internal = NULL)
}
\arguments{
\item{wf_mapping}{Data frame created by setup_which_factors}

\item{wf_internal}{Character vector of internal wf_* column names}
}
\value{
Named character vector of original column names
}
\description{
Retrieves original column names for reference
}
