% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe-result-class.R
\name{new_universe_result}
\alias{new_universe_result}
\title{Build a universe result (internal helper)}
\usage{
new_universe_result(b, ci.lb, ci.ub, pval)
}
\arguments{
\item{b, ci.lb, ci.ub, pval}{Numeric scalars.}
}
\value{
An object of class \code{c("universe_result","list")}.
}
\description{
Build a universe result (internal helper)
}
\keyword{internal}
