% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_descriptive_spec_curve.R
\name{plotly_descriptive_spec_curve}
\alias{plotly_descriptive_spec_curve}
\title{Interactive Specification Curve Plot}
\usage{
plotly_descriptive_spec_curve(
  data,
  ylim_lower = NULL,
  ylim_upper = NULL,
  colorblind_friendly = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{data}{Data frame containing the specification data and results.}

\item{ylim_lower}{Numeric. Lower limit for the y-axis of the forest plot.}

\item{ylim_upper}{Numeric. Upper limit for the y-axis of the forest plot.}

\item{colorblind_friendly}{Logical. Use a colorblind-friendly palette (default: TRUE).}

\item{interactive}{Logical. Whether to create an interactive Plotly plot (`TRUE`) or static ggplot plot (`FALSE`). Default is `TRUE`.}
}
\value{
A `plotly` object containing the interactive specification curve plot if `interactive = TRUE`, or a combined `ggplot` object if `interactive = FALSE`.
}
\description{
Creates an interactive specification curve plot with a forest plot for effect sizes and a tile plot for specification factors.
}
\details{
Combines a forest plot for effect sizes with confidence intervals and a tile plot visualizing specification factors.
}
\examples{
plotly_descriptive_spec_curve(data, ylim_lower = -0.5, ylim_upper = 1.5, interactive = TRUE)

}
