% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_ma_method.R
\name{register_ma_method}
\alias{register_ma_method}
\title{Register a meta-analysis estimator}
\usage{
register_ma_method(
  name,
  fun,
  dependencies = c("select_min", "select_max", "aggregate", "modeled")
)
}
\arguments{
\item{name}{Character scalar (key used in `ma_methods`)}

\item{fun}{Function(dat) → list(b, ci.lb, ci.ub, pval)}

\item{dependencies}{Character vector: "select_min", "select_max", "aggregate", "modeled"}
}
\description{
Register a meta-analysis estimator
}
\keyword{internal}
